+++ packages/fdc3-standard/src/context/ContextType.ts
@@ -39,7 +39,12 @@ export type ExperimentalContextType =
   | 'fdc3.orderList'
   | 'fdc3.product'
   | 'fdc3.trade'
-  | 'fdc3.tradeList';
+  | 'fdc3.tradeList'
+  | 'fdc3.security.user'
+  | 'fdc3.security.userRequest'
+  | 'fdc3.security.symmetricKeyRequest'
+  | 'fdc3.security.symmetricKeyResponse'
+  | 'fdc3.security.encryptedContext';
 
 /**
  * @see https://fdc3.finos.org/docs/context/spec
@@ -78,4 +83,9 @@ export enum ContextTypes {
   TradeList = 'fdc3.tradeList',
   TransactionResult = 'fdc3.transactionResult',
   Valuation = 'fdc3.valuation',
+  SecurityUser = 'fdc3.security.user',
+  SecurityUserRequest = 'fdc3.security.userRequest',
+  SecuritySymmetricKeyRequest = 'fdc3.security.symmetricKeyRequest',
+  SecuritySymmetricKeyResponse = 'fdc3.security.symmetricKeyResponse',
+  SecurityEncryptedContext = 'fdc3.security.encryptedContext',
 }